requirements.txt
pandas
scikit-learn
python-dotenv
streamlit 
streamlit-aggrid
matplotlib
seaborn
altair
openai